requirements.txt
biopython
biopython bcbio-gff